+++ packages/popmotion/src/animations/spring/index.ts
@@ -4,7 +4,7 @@ import action from '../../action';
 import { Action } from '../../action';
 import vectorAction from '../../action/vector';
 import { speedPerSecond } from '../../calc';
-import onFrame from '../on-frame';
+import onFrame from '../every-frame';
 import { SpringInterface, SpringProps } from './types';
 
 const spring = ({